PrjBDF/producers/requirements.txt
praw
kafka-python-ng
google-api-python-client
requests